requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.0.331
langsmith==0.0.60
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
parso==0.8.3
pinecone-client==2.2.4
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
spicy==0.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
win32-setctime==1.1.0
yarl==1.9.2
